pyproject.toml
[project]
name = "rag-chatbot-project"
version = "0.1.0"
description = "FAQ Chatbot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.7.0",
    "beautifulsoup4>=4.13.4",
    "chromadb>=1.0.12",
    "datasets>=3.6.0",
    "exif>=1.6.1",
    "exifread>=3.3.1",
    "fastapi>=0.115.9",
    "fpdf>=1.7.2",
    "gradio>=5.33.1",
    "hachoir>=3.3.0",
    "huggingface-hub[hf-xet]>=0.33.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.25",
    "langchain-ollama>=0.3.3",
    "lxml>=5.4.0",
    "markdown>=3.8",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.0",
    "pathlib>=1.0.1",
    "piexif>=1.1.3",
    "pillow>=11.2.1",
    "plotly>=6.1.2",
    "pyexiftool>=0.5.6",
    "pymongo>=4.13.0",
    "pypdf>=5.6.0",
    "pytesseract>=0.3.13",
    "scikit-learn>=1.7.0",
    "torch>=2.7.1",
    "transformers>=4.52.4",
    "unstructured>=0.17.2",
    "uvicorn>=0.34.3",
]
